Update strawberry.py to start a transaction instead of a span
Operations in Grapqhl are logically equivalent units to Views in HTTP - and should use a transaction instead of a span as the parent object. They represent a single activity instance we want to measure or track.

Making them a transaction enables us to have visibility of them at the Performance metric tab and see which operations are simply causing a high load.